requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=13.0.0
unstructured[pdf]>=0.10.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Evaluation
scikit-learn>=1.3.0
rouge-score>=0.1.2
bert-score>=0.3.13

# NLP utilities
spacy>=3.7.0
nltk>=3.8.1

# Demo
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
pytest>=7.4.0
